requirements.txt
# pip requirements
ansibleguy-nftables